Crate frame_remote_externalities

Remote Externalities

An equivalent of sp_io::TestExternalities that can load its state from a remote substrate based chain, or a local state snapshot file.

Structs

Builder

Builder for remote-externalities.

OfflineConfig

Configuration of the offline execution.

OnlineConfig

Configuration of the online execution.

RemoteExternalities

An externalities that acts exactly the same as sp_io::TestExternalities but has a few extra bits and pieces to it, and can be loaded remotely.

SnapshotConfig

Configuration of the state snapshot.

Enums

Mode

The execution mode.

Transport

Description of the transport protocol (for online execution).